+++ rust/candid/src/bindings/candid.rs
@@ -107,7 +107,7 @@ pub fn pp_ty(ty: &Type) -> RcDoc {
         Func(ref func) => kwd("func").append(pp_function(func)),
         Service(ref serv) => kwd("service").append(pp_service(serv)),
         Class(ref args, ref t) => {
-            let doc = pp_args(&args).append(" ->").append(RcDoc::space());
+            let doc = pp_args(args).append(" ->").append(RcDoc::space());
             match t.as_ref() {
                 Service(ref serv) => doc.append(pp_service(serv)),
                 Var(ref s) => doc.append(s),