+++ rust/candid/tests/assets/ok/fieldnat.d.ts
@@ -1,7 +1,7 @@
 import type { Principal } from '@dfinity/principal';
-export interface non_tuple { _1_ : string, _2_ : string };
+export interface non_tuple { _1_ : string, _2_ : string }
 export type tuple = [string, string];
-export default interface _SERVICE {
+export interface _SERVICE {
   'bab' : (arg_0: bigint, arg_1: bigint) => Promise<undefined>,
   'bar' : (arg_0: { '2' : bigint }) => Promise<undefined>,
   'bas' : (arg_0: [bigint, bigint]) => Promise<[string, bigint]>,
@@ -9,4 +9,4 @@ export default interface _SERVICE {
   'bba' : (arg_0: tuple) => Promise<non_tuple>,
   'bib' : (arg_0: [bigint]) => Promise<{ _0_ : bigint }>,
   'foo' : (arg_0: { _2_ : bigint }) => Promise<{ _2_ : bigint, '_2' : bigint }>,
-};
+}